+++ docs/versionhistory.rst
@@ -6,6 +6,8 @@ This library adheres to `Semantic Versioning 2.0 <https://semver.org/#semantic-v
 **UNRELEASED**
 
 - Added ``InstrumentationWarning`` to the public API
+- Changed ``@typechecked`` to skip instrumentation in optimized mode, as in typeguard
+  2.x
 - Avoid type checks where the types in question are shadowed by local variables
 - Fixed instrumentation using ``typing.Optional`` without a subscript when the subscript
   value was erased due to being an ignored import